+++ CHANGELOG.md
@@ -91,6 +91,7 @@
 - CONTRACT : Set Invoice date with newly added options for invoicing Moment
 - Inventory : Added new field 'typeSelect' and change in product doman filter.
 - Address: Street have now a dedicated object.
+- PRODUCTION ORDER : user can define manuf order's planned end date while creating production order from product form
 
 ## Improvements
 - Contract: added button to manually close contract if termination date was set in the future